Renaming, moving TODOs to notes out of git
